backend/requirements.txt
# Meta Ads → Supabase Pipeline
# Install with: pip install -r requirements.txt

# HTTP requests — Meta API calls
requests==2.31.0

# PostgreSQL driver — Supabase direct DB connection
psycopg2-binary==2.9.9

# Environment variables — loads .env file
python-dotenv==1.0.0

# Job scheduler — hourly + daily cron in scheduler.py
schedule==1.2.